+++ src/polymer/databind-with-unknown-property.ts
@@ -14,7 +14,7 @@
 
 import * as dom5 from 'dom5';
 import {Document, isPositionInsideRange, ParsedHtmlDocument, Severity, SourceRange, Warning} from 'polymer-analyzer';
-import {DatabindingExpression} from 'polymer-analyzer/lib/polymer/expression-scanner';
+import {AttributeDatabindingExpression, DatabindingExpression} from 'polymer-analyzer/lib/polymer/expression-scanner';
 
 import {HtmlRule} from '../html/rule';
 import {sharedProperties} from '../html/util';
@@ -95,7 +95,6 @@ export class DatabindWithUnknownProperty extends HtmlRule {
         }
       }
 
-
       // TODO(rictic): also validate computed properties and observers once
       //     https://github.com/Polymer/polymer-analyzer/pull/552 has
       //     landed.
@@ -122,9 +121,9 @@ export class DatabindWithUnknownProperty extends HtmlRule {
           // text node), if it's a bidirectional binding, and it's the only
           // property in the expression. (this isn't perfect, it misses some
           // strange stuff with literals like foo(10, 20) but it's good enough.)
-          return use.expression.databindingInto === 'attribute' &&
+          return use.expression instanceof AttributeDatabindingExpression &&
               use.expression.direction === '{' &&
-              use.expression.properties.length === 1;
+              use.expression.isCompleteBinding;
         });
         // TODO(rictic): when we add the ability to configure a lint pass we
         //     should allow users to force all properties to be declared.