+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
    Unreleased section, uncommenting the header as necessary.
 -->
 
-<!--## Unreleased-->
+## Unreleased
+
+* Populate `astNode` on `PolymerElementMixin`.
 
 ## [2.0.0-alpha.32] - 2017-03-13